requirements.txt
asgiref==3.2.7
Django==3.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
